[Rename] 주소 수정 멤버 페이지 user -> member로 변경했습니당
