+++ src/App.js
@@ -8,8 +8,8 @@ import Home from './pages/Home';
 import Login from './pages/Login';
 import Signup from './pages/Signup';
 import EventDetails from './pages/EventDetail';
-import User from './pages/User';
-import UserDetail from './pages/UserDetail';
+import Member from './pages/Member';
+import MemberDetail from './pages/MemberDetail';
 
 import { ThemeProvider } from 'styled-components';
 import theme from './styles/theme';
@@ -24,8 +24,8 @@ function App() {
         <Route path="/home" element={<Home />} />
         <Route path="/signup" element={<Signup />} />
         <Route path="/event" element={<EventDetails />} />
-        <Route path="/user" element={<User />} />
-        <Route path="/user/detail" element={<UserDetail />} />
+        <Route path="/member" element={<Member />} />
+        <Route path="/member/detail" element={<MemberDetail />} />
       </Routes>
     </ThemeProvider>
   );
